refactor(models): consolidate bank account validation logic into shared concern
Remove 900+ lines of duplicate code across 18 bank account models by creating
a reusable BankAccountValidations concern that provides:

- Shared validation methods for bank_code, branch_code, and account_number
- Common helper methods: routing_number, account_number_visual, to_hash
- Support for different validation patterns (regex vs IBAN)
- Flexible routing number formats (concatenation vs dash-separated)

Refactored models:
- ChileBankAccount, ColombiaBankAccount, BoliviaBankAccount
- BangladeshBankAccount, JamaicaBankAccount, BruneiBankAccount
- ArmeniaBankAccount, NigeriaBankAccount, MalaysiaBankAccount
- IndonesiaBankAccount, PhilippinesBankAccount, ThailandBankAccount
- VietnamBankAccount, SingaporeanBankAccount, IsraelBankAccount
- KuwaitBankAccount, JapanBankAccount, KoreaBankAccount

All tests passing. Backwards compatible with existing functionality.
