+++ app/models/armenia_bank_account.rb
@@ -1,16 +1,12 @@
 # frozen_string_literal: true
 
 class ArmeniaBankAccount < BankAccount
-  BANK_ACCOUNT_TYPE = "AM"
+  include BankAccountValidations
 
+  BANK_ACCOUNT_TYPE = "AM"
   BANK_CODE_FORMAT_REGEX = /^[0-9A-Za-z]{8,11}$/
   ACCOUNT_NUMBER_FORMAT_REGEX = /^\d{11,16}$/
   private_constant :BANK_CODE_FORMAT_REGEX, :ACCOUNT_NUMBER_FORMAT_REGEX
-
-  alias_attribute :bank_code, :bank_number
-
-  validate :validate_bank_code
-  validate :validate_account_number
 
   def routing_number
     bank_code.to_s
@@ -27,27 +23,4 @@ class ArmeniaBankAccount < BankAccount
   def currency
     Currency::AMD
   end
-
-  def account_number_visual
-    "******#{account_number_last_four}"
-  end
-
-  def to_hash
-    {
-      routing_number:,
-      account_number: account_number_visual,
-      bank_account_type:
-    }
-  end
-
-  private
-    def validate_bank_code
-      return if BANK_CODE_FORMAT_REGEX.match?(bank_code)
-      errors.add :base, "The bank code is invalid."
-    end
-
-    def validate_account_number
-      return if ACCOUNT_NUMBER_FORMAT_REGEX.match?(account_number_decrypted)
-      errors.add :base, "The account number is invalid."
-    end
 end
