+++ app/models/brunei_bank_account.rb
@@ -1,22 +1,12 @@
 # frozen_string_literal: true
 
 class BruneiBankAccount < BankAccount
+  include BankAccountValidations
+
   BANK_ACCOUNT_TYPE = "BN"
-
   BANK_CODE_FORMAT_REGEX = /^[0-9a-zA-Z]{8,11}$/
-  private_constant :BANK_CODE_FORMAT_REGEX
-
   ACCOUNT_NUMBER_FORMAT_REGEX = /^[0-9]{1,13}$/
-  private_constant :ACCOUNT_NUMBER_FORMAT_REGEX
-
-  alias_attribute :bank_code, :bank_number
-
-  validate :validate_bank_code
-  validate :validate_account_number
-
-  def routing_number
-    "#{bank_code}"
-  end
+  private_constant :BANK_CODE_FORMAT_REGEX, :ACCOUNT_NUMBER_FORMAT_REGEX
 
   def bank_account_type
     BANK_ACCOUNT_TYPE
@@ -29,27 +19,4 @@ class BruneiBankAccount < BankAccount
   def currency
     Currency::BND
   end
-
-  def account_number_visual
-    "******#{account_number_last_four}"
-  end
-
-  def to_hash
-    {
-      routing_number:,
-      account_number: account_number_visual,
-      bank_account_type:
-    }
-  end
-
-  private
-    def validate_bank_code
-      return if BANK_CODE_FORMAT_REGEX.match?(bank_code)
-      errors.add :base, "The bank code is invalid."
-    end
-
-    def validate_account_number
-      return if ACCOUNT_NUMBER_FORMAT_REGEX.match?(account_number_decrypted)
-      errors.add :base, "The account number is invalid."
-    end
 end
