+++ app/models/israel_bank_account.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class IsraelBankAccount < BankAccount
+  include BankAccountValidations
+
   BANK_ACCOUNT_TYPE = "IL"
 
   validate :validate_account_number, if: -> { Rails.env.production? }
@@ -20,18 +22,4 @@ class IsraelBankAccount < BankAccount
   def account_number_visual
     "#{country}******#{account_number_last_four}"
   end
-
-  def to_hash
-    {
-      account_number: account_number_visual,
-      bank_account_type:
-    }
-  end
-
-  private
-    def validate_account_number
-      return if Ibandit::IBAN.new(account_number_decrypted).valid?
-
-      errors.add :base, "The account number is invalid."
-    end
 end
