+++ app/models/nigeria_bank_account.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class NigeriaBankAccount < BankAccount
+  include BankAccountValidations
+
   BANK_ACCOUNT_TYPE = "NG"
 
   BANK_CODE_FORMAT_REGEX = /^([0-9a-zA-Z]){8,11}$/
@@ -9,14 +11,7 @@ class NigeriaBankAccount < BankAccount
   ACCOUNT_NUMBER_FORMAT_REGEX = /^\d{10}$/
   private_constant :ACCOUNT_NUMBER_FORMAT_REGEX
 
-  alias_attribute :bank_code, :bank_number
-
-  validate :validate_bank_code
   validate :validate_account_number, if: -> { Rails.env.production? }
-
-  def routing_number
-    "#{bank_code}"
-  end
 
   def bank_account_type
     BANK_ACCOUNT_TYPE
@@ -33,23 +28,4 @@ class NigeriaBankAccount < BankAccount
   def account_number_visual
     "#{country}******#{account_number_last_four}"
   end
-
-  def to_hash
-    {
-      routing_number:,
-      account_number: account_number_visual,
-      bank_account_type:
-    }
-  end
-
-  private
-    def validate_bank_code
-      return if BANK_CODE_FORMAT_REGEX.match?(bank_code)
-      errors.add :base, "The bank code is invalid."
-    end
-
-    def validate_account_number
-      return if ACCOUNT_NUMBER_FORMAT_REGEX.match?(account_number_decrypted)
-      errors.add :base, "The account number is invalid."
-    end
 end
