+++ app/models/philippines_bank_account.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class PhilippinesBankAccount < BankAccount
+  include BankAccountValidations
+
   BANK_ACCOUNT_TYPE = "PH"
 
   BANK_CODE_FORMAT_REGEX = /\A[A-Za-z0-9]{8,11}\z/
@@ -8,15 +10,6 @@ class PhilippinesBankAccount < BankAccount
 
   ACCOUNT_NUMBER_FORMAT_REGEX = /\A[0-9]{1,17}\z/
   private_constant :ACCOUNT_NUMBER_FORMAT_REGEX
-
-  alias_attribute :bank_code, :bank_number
-
-  validate :validate_bank_code
-  validate :validate_account_number
-
-  def routing_number
-    "#{bank_code}"
-  end
 
   def bank_account_type
     BANK_ACCOUNT_TYPE
@@ -29,27 +22,4 @@ class PhilippinesBankAccount < BankAccount
   def currency
     Currency::PHP
   end
-
-  def account_number_visual
-    "******#{account_number_last_four}"
-  end
-
-  def to_hash
-    {
-      routing_number:,
-      account_number: account_number_visual,
-      bank_account_type:
-    }
-  end
-
-  private
-    def validate_bank_code
-      return if BANK_CODE_FORMAT_REGEX.match?(bank_code)
-      errors.add :base, "The bank code is invalid."
-    end
-
-    def validate_account_number
-      return if ACCOUNT_NUMBER_FORMAT_REGEX.match?(account_number_decrypted)
-      errors.add :base, "The account number is invalid."
-    end
 end
